fix alignment problem and random generator.
git-svn-id: file:///svn/nsd/trunk@3682 a26ef69c-88ff-0310-839f-98b793d9c207

+++ util.c
@@ -852,9 +852,9 @@ random_generate(int max)
 #ifdef HAVE_ARC4RANDOM_UNIFORM
     return (int) arc4random_uniform(max);
 #elif HAVE_ARC4RANDOM
-    return (int) arc4random() % max;
+    return (int) (arc4random() % max);
 #else
-    return (int) random() % max;
+    return (int) ((unsigned)random() % max);
 #endif
 }
